requirements.txt
Pyrogram
TgCrypto
requests
selenium
python-dotenv
